chore: retract v1 (#1325)
`pkg/**` and `Makefile` were removed. 
I kept go.mod for the release. We will remove it later.

Dummy CI was added for non-code changes to files in the root.
Changes in linter config will trigger v2 and executions workflows.

I plan to merge it and then tag it with v1.67.5