fix: use waitgroup to wait for routines in test
The hookCompleted channel was never read from, only written to.
The test also blocked when writing the second time to that channel, until test timeout, when run in isolation.
It mistankingly was displayed as successfully finished when run via
go test ./... because the timeout there is different and then the internal timeout of the test was faster and mistankingly marked the test as successfully finished.

This got fixed by using a waitgroup, which is a better fit here. It prevents blocking the routines, and we now actually wait for the routines to finish or we fail.

Also the heartbeat message is not "{}" but "heartbeat". This got fixed as well